python/requirements.txt
gymnasium>=0.29.0
stable-baselines3>=2.0.0
pyzmq>=25.0.0
numpy>=1.24.0
customtkinter>=5.0.0
tensorboard>=2.14.0
pytest>=7.0.0
